requirements.txt
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0.1
pathlib>=1.0.1 